requirements.txt
streamlit
pandas
Pillow
python-dotenv
openai
pinecone-client
replicate
requests